* Moved logger hook to more global location.
